refactor: move files to src folder update the import paths accordingly
