+++ src/RSS_abstracts/RSSEntry.py
